moved HTTPNetworkingError to its own file and renamed test file name
